pyproject.toml
[project]
name = "cp-font-gen"
version = "1.0.0"
description = "Generate minimal bitmap fonts for CircuitPython devices"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "cp-font-gen contributors"}
]
keywords = [
    "circuitpython",
    "fonts",
    "embedded",
    "bitmap",
    "pcf",
    "bdf",
    "adafruit",
    "microcontroller"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Embedded Systems",
    "Topic :: Multimedia :: Graphics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]

dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0",
    "fonttools>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/graybear-io/cp-font-gen"
Documentation = "https://github.com/graybear-io/cp-font-gen/tree/main/docs"
Repository = "https://github.com/graybear-io/cp-font-gen"
Issues = "https://github.com/graybear-io/cp-font-gen/issues"

[project.scripts]
cp-font-gen = "cp_font_gen.cli:cli"

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.8.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "types-pyyaml>=6.0.12.20250915",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cp_font_gen"]

# Ruff configuration
[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# Mypy configuration
[tool.mypy]
python_version = "3.10"
warn_return_any = false  # Permissive for existing code
warn_unused_configs = true
disallow_untyped_defs = false  # Start permissive, can tighten later
check_untyped_defs = true
ignore_missing_imports = true  # Permissive for third-party libraries
no_implicit_optional = false  # Allow existing patterns

# Ignore specific third-party libraries without stubs
[[tool.mypy.overrides]]
module = ["fontTools.*"]
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v"
